+++ test/manual/handlebars/src/views/helpers/bold.js
@@ -1,0 +1,2 @@
+const Handlebars = require('handlebars');
+module.exports = (options) => new Handlebars.SafeString(`<strong>${options.fn(this)}</strong>`);